# docs/swagger.yaml
definitions:
...
  entity.Page:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
  entity.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: integer
      roles:
        items:
          type: string
        type: array
      type:
        type: string
      updatedAt:
        type: string
    type: object
  entity.UserChangePassword:
    properties:
      id:
        type: string
      name:
        type: string
      original:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/user:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: ""
      tags:
      - 删除用户
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      tags:
      - 根据用户名获取用户对象
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      tags:
      - 更新用户
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      tags:
      - 创建用户
  /api/v1/user/:user_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
      tags:
      - 根据用户ID获取用户对象
  /api/v1/user/password:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UserChangePassword'
      responses:
        "200":
          description: ""
      tags:
      - 修改用户密码
  /api/v1/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        required: true
        type: string
      - description: 长度
        in: query
        name: size
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Page'
      tags:
      - 根据用户ID获取用户对象
swagger: "2.0"